[panel 2] - fix poster on left
Saffron won the magic challenge. Sometime, I keep this idea so long in mind I forgot to explain them better. The poster now has more details: the arena with the stripped colored balloon in background, the public, and a more obvious podium. Little fix were added here and there for Spirulina and Camomille.